refactor: pull video views into file
